% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{DiffAedes}
\alias{DiffAedes}
\title{Aedes aegypti RNA-seq differential expression data example}
\format{A data.frame with 1963 rows and 3 variables}
\description{
A sample fragment of differential gene expression from an
RNA-seq experiment of Aedes aegypti mosquito.
}
\examples{
data(DiffAedes)
}
\references{
AKBARI, O. S. et al. The developmental transcriptome of the
mosquito aedes aegypti, an invasive species and major arbovirus vector.
G3: Genes— Genomes— Genetics, Genetics Society of America, v. 3, n. 9,
p. 1493–1509, 2013.
}
