% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ExpressionAedes}
\alias{ExpressionAedes}
\title{Aedes aegypti RNA-seq data expression}
\format{A data.frame with 2000 rows and 5 variables}
\description{
A sample fragment of gene expression from an RNA-seq experiment
of Aedes aegypti mosquito.
}
\references{
AKBARI, O. S. et al. The developmental transcriptome of the
mosquito aedes aegypti, an invasive species and major arbovirus vector. G3:
Genes— Genomes— Genetics, Genetics Society of America, v. 3, n. 9,
p. 1493–1509, 2013.
#' @examples
data(ExpressionAedes)
}
