% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{GeneFunctionAedes}
\alias{GeneFunctionAedes}
\title{Relation between Aedes aegypti genes and KEGG pathways}
\format{A data.frame with 782 rows and 2 variables}
\description{
A relation between the GFAGs present in ResultAnalysisAedes and
their respective genes.
}
\examples{
data(GeneFunctionAedes)
}
\references{
Molan, A. L. 2018. “Construction of a Tool for Multispecies Genic
Functional Enrichment Analysis Among Comparative Samples.” Master’s thesis,
Institute of Biosciences of Botucatu – Univ. Estadual Paulista.
http://hdl.handle.net/11449/157105.
}
