% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ResultAnalysisAedes}
\alias{ResultAnalysisAedes}
\title{Result from an example of ADAM Aedes aegypti analysis}
\format{A data.frame with 87 rows and 22 variables}
\description{
Result from ADAM Aedes aegypti analysis according to
ExpressionAedes data.
}
\examples{
data(ResultAnalysisAedes)
}
\references{
Molan, A. L. 2018. “Construction of a Tool for Multispecies Genic
Functional Enrichment Analysis Among Comparative Samples.” Master’s thesis,
Institute of Biosciences of Botucatu – Univ. Estadual Paulista.
http://hdl.handle.net/11449/157105.
}
