% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFAGtargetUi.R
\name{GFAGtargetUi}
\alias{GFAGtargetUi}
\title{GFAGtargetUi}
\usage{
GFAGtargetUi(browser)
}
\arguments{
\item{browser}{is a logical variable necessary to run the app.
When browser=TRUE the app is launched in your default web browser.
When browser=FALSE the app is launched in your local machine.}
}
\value{
return a shiny Graphical User Interface (GUI).
}
\description{
Launch Launch GFAGtargetUi Graphical User Interface (GUI) in
local machine or default browser
}
\details{
This shiny (GUI) enables the creation of a plot that shows the
expression of the targets in a selected path between control and case.
See the vignette for more information.
}
\examples{
#Creating input files
data(DiffHs)
data(ExpressionHs)
data(GeneFunctionHs)
data(ResultAnalysisHs)
\dontrun{
#Input file 1
write.table(DiffHs,file = "DiffHs.txt",sep = "\\t", col.names = TRUE,
row.names = FALSE, quote = FALSE)
#Input file 2
write.table(ExpressionHs,file = "ExpressionHs.txt",sep = "\\t",
col.names = TRUE, row.names = FALSE, quote = FALSE)
#Input file 3
write.table(GeneFunctionHs,file = "GeneFunctionHs.txt",sep = "\\t",
col.names = TRUE, row.names = FALSE, quote = FALSE)
#Input file 4
write.table(ResultAnalysisHs,file = "ResultAnalysisHs.txt",sep = "\\t",
col.names = TRUE, row.names = FALSE, quote = FALSE)
#Grahphical analysis
GFAGtargetUi(browser = TRUE) # Launch the app in your default web browser.
GFAGtargetUi(browser = FALSE) # Launch the app in your local machine.
}
}
\author{
Giordano Bruno Sanches Seco <giordano.bruno@unesp.br>
}
